requirements.txt
Flask
qrcode[pil]
Pillow
